requirements.txt
fastapi[standard]
uvicorn[standard]
httpx
SQLAlchemy
psycopg2-binary
bcrypt
pydantic
fastapi_mail
redis
phonenumbers
PyJWT
geopy